backend/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
oauthlib==3.2.2
packaging==24.2
pyasn1==0.6.1
pymongo==4.5.0
python-dotenv==1.0.0
python-jose==3.3.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
six==1.16.0
textblob==0.17.1
tqdm==4.67.0
tweepy==4.14.0
urllib3==2.2.3
Werkzeug==3.1.3
